// apps/dashboard/src/app-deps/(clean-code)/(sales)/types.ts
export interface AddressBookMeta {
    zip_code;
    lat?: number | null;
    lng?: number | null;
    placeId?: string;
    placeSearchText?: string;
}
export type CustomerMeta = {
    netTerm?: string;
};